flight-bot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pymupdf==1.24.10
Pillow==10.4.0
reportlab==4.2.2
python-multipart==0.0.9
